add genderize
rewrite the terminal version of Genderize to work on local server; add start message; move request() func in module file; feat: better name capitalization in response message
