+++ miamoremustachio/genderize/main.js
@@ -1,50 +1,31 @@
 const {
-    API,
-    EXIT_COMMAND,
-    getResponseError,
+    START_MESSAGE,
     getNameError,
     getCapitalizedName,
-    readline,
+    request,
 } = require('./modules');
 
+const http = require('http');
+const server = http.createServer().listen(3000);
 
-function request(name) {
-    const url = `${API}?name=${name}`;
-    
-    return fetch(url)
-    .then(response => {
-        if (!response.ok) {
-            throw new Error(getResponseError(response));
+server.on('request', (req, res) => {
+    if (req.url === '/') {
+        return res.end(START_MESSAGE);
+    };
+
+    const nameFromUrl = req.url.slice(1);
+
+    request(nameFromUrl)
+    .then(data => {
+        const name = getCapitalizedName(data.name);
+        const gender = data.gender;
+        const probability = (Math.trunc(data.probability * 100));
+
+        if (!gender) {
+            throw new Error(getNameError(name));
         };
-
-        return response.json();
+        
+        res.end(`The name ${name} is ${gender} with a probability of ${probability}%`);
     })
-}
-
-function askName() {
-    readline.question("What is your name? ", name => {
-        if (name === EXIT_COMMAND) {
-            return readline.close();
-        };
-
-        request(name)
-        .then(data => {
-            const name = getCapitalizedName(data.name);
-            const gender = data.gender;
-            const probability = (Math.trunc(data.probability * 100));
-
-            if (!gender) {
-                throw new Error(getNameError(name));
-            };
-            
-            console.log(`The name ${name} is ${gender} with a probability of ${probability}%`);
-        })
-        .catch(error => console.error(error.message))
-        .finally(() => {
-            console.log(`* (type <${EXIT_COMMAND}> to exit)`);
-            askName();
-        });
-    });
-}
-
-askName();
+    .catch(error => res.end(error.message));
+});
